Codes to rasterize and prepare SIMCI data
This commit includes new scripts in R and Python to process SIMCI
shapefiles to raster data. Since R is taking a lot of time to load using
rgdal bindings, I use Python Geopandas and RasterIO to load the data and
rasterize using nightlight data as a template. The function rasterize is
created and save in source folder. This function takes a variable from
the shapefile and rasterize its contents

TODO
 - Loop rasterize function to all numerical columns in shapefiles
 - Solve GDAL_DATA CRS environment variable / hotfix

BUGS
 - GDAL can't find projection data .csv