Data extraction from SIMCI rasters
This commit includes R scripts to read Python processed rasters and
extract their features. To do this, we read the data and resample it in
order to have comparable observations across datasets (i.e.
deforestation, nightlights, and covariates). In the raster_simci.R we
leave a piece of R code as an example of how to run the rasterization in
R instead of Python.

TODO:

 - Set up folders with global environment variables