Changes to cleaning scripts: split code and source functions
This new commit splits the cleaning of borders and distance calculation
between to new files, one for parks and other one for territories. This
makes easier the process and avoid the overloading of the GlobalEnv in R
with a lot of heavy objects. Also, the functions are sourced from ./R
and the code is commented.