New structured code to clean bad-bad treatments/controls
New source code is generated in ./R to manage functions needed to clean
all borders in treatment territories. A new script
colombia_continental.R is created to avoid repetition when calling
Colombia's geometry, now this script can be sourced and Colombia's
shapes are loaded. In general, cleaning scripts are more readable and
clean.