New fancy script to define enviromental variables in R!
Since R do not source enviromental variables from .bash_profile (.bashrc
in the Linux case), this script creates enviromental variables and paste
them directly in .Renviron file in R bin/local files. Another possible
approach is create a .Renviron in the project folder and source it
everytime we run R. That can be a better options if replication is done
in other boxes.