Changes to R functions to calculate distances to clean borders
This commit includes changes to distance calculation functions. To make
things easier two functions are created, one for natural parks, and a
second one to black and indigenous territories. This approach is far
from perfect, but for now is convenient. We need to create a new
function with more general specs and avoid hard coding arguments into
the function.