Updated clean geometries and relative paths
This new scripts use parks and territories up to 2016 (the same update
of Hansen data) to clean borders and calculate distances. New
corrections are made. For instance, we use the data path from the
enviroment variable (the use of setwd() is deprecated).

Functions and individual scripts are sourced from ./R, and a more clean
and commented version is commited.