Changes/Simplification of environment variables
The previous environment_variables script modified the .Renviron file in
the R installation files. This was a good approach, but if the code was
replicated in a box with different OS, it was prone to error. This new
environment_variables scrip creates a new .Renviron in the local
directory. R will read this new local file and the env_vars can be
easily used.