Update distance datasets for all protected areas
Changes in the folder structure and expanded loops with a new date span
to include new deforestation data (2001-2016). This code, as the
previous one, creates .rds files for each protected area and year to
identify the neares effective border to the pixel. This new changes
includes the use of enviroment variables plus some minor changes in the
identifiers.