Changes in dataframe creation for Strategy 2
This new code changes the loops that were created to build year/area
databases. This new loops copy the former ones, but create lists with
year and type area, named by the type, for easy indexing. This allows to
have a better understanding of what's happening under the hood of the
code.

Please do note that this changes will only apply to Strategy 2.