New covariates: clumps for 5km and 10km
To solve estimation problems in areas with low population/light, we
create new covariates: clumps within a 5km and 10 km influence zone. The
past code used the clump raster function to identify cells with values
inside the raster. We created  a shape file geometry from the clump, and
apply a buffer to design this influence zones.

Cells are identified inside the "res[[1]]" deforestation template raster
and saved as csv as [1,0] dummy variable