New table function: add date and change column str
Changes in function from rd_obj to df. Now the function reads the file
directly using readRDS internally. The idea behind this is to extract
the date of the model. This will allow us to have more control/features
over the different iterations of the modeling (a la DSaPP). Also, since
this is exported to LaTeX, integers now have thousands divisors.