Add functions to estimate RD test + tables
This commit adds new functionality to the RD tests framework. First a
permutation test is added to use the RATest function and apply it to a
list of data.frames. In the same fashion,the new tables function
(rd_to_df2) function allow to take a list of files, lists, or individual
object and process it to a df. This functionality is made to allow any
time of object into the function and to allow flexibility in the inputs.

Also, a placebo function is created with hardocded values for the
covariates. This is non-optimal, but for the actual use is not a big
contraint.