Permutation tests
This commit add a tests directory with test scripts. The
permutation_tests.R call the perm_test_list function in R/ and apply it
to a list of dataframes. Also, permutation_test.R calculates
descriptives for a defined bandwidth and concatenate its results with
the test results. The output can be exported to a LaTeX table easily.