New merging script: simplify tasks!
A merge script is created to be called everytime a modeling function is
needed. This create simple modeling files and more simple calls to data
functions. Some functionality can be lost, but a simple source() to
modeling/merge_datasets.R can solve that.

Also! New script for summary tables and graphs!

BUGS:
- Some files aren't modified yet.