Changes to modeling scripts
This new changes in modeling create a more complete a readable modeling.
The models here are the esential ones (needed for each table of the
paper) and they save the results as .rds to the results dir. The final
approach would be to use reg_dis.R as the only modeling script, but some
iterations of the reg_dis_2.R are still needed.

The modeling reg_dis.R file needs to source the merge scripts!