Correction in iteration for sensibility test
CHange of strategy! The past code was creating a cartesian product of
all elements in a set of list creating incorrect permutation and
creating huge and unnecessary models. This new approach is cleaner.
First, we create a data frame with all the necessary data to iterate
(including the data frames! thanks tibble!) and pmap using the same
function we used in the past.

This is the best way to achieve things in life and would be good to use
the same parameter approach in the modeling part. I guess those changes
will come later