Plot sensibility estimates in RD
This commit uses pmap to map the parameters dataframe into the
sensibility rd function. The pmap go iteratively for each row of the df
and use the columns as parameters for estimation. At the end, R will
output a list of results (30 rows per parameter row: 1 to 15 by 0.5).

This lambda approach helps to make results easier to read and makes the
estimation more clean that the placebo ones, where we have to create
vectors of names and variables to identify the estimations.